% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{readPng}
\alias{readPng}
\title{Read a PNG file}
\usage{
readPng(file)
}
\arguments{
\item{file}{A character string giving the file name to read from.}
}
\value{
An array of class \code{"loder"}.
}
\description{
Read an image from a PNG file and convert the pixel data into an R array.
}
